day 1 part 2 solution
Definitely made it harder than it had to be by pipelining inputs in the RTL instead of unpacking in the driver, but that also made it more fun/worthwhile imo, so glad I did it that way.

That said, I had one too few valid stages, which meant the answer was off by one, and this was particularly hard to track down!

One last thing: there's far more type annotations that I'd like; that's another way in which kaze really sucks rn.