day 1: Initial Parser module impl
Something didn't quite feel right about not parsing in RTL as well, and it didn't seem that difficult, so I gave it a go. I don't have time right now to clean it up or hook it up to the part solutions, so I'll do that later, but for now, at least it's in a working state (I think; I also didn't test backpressure cases yet!).